Replace shortcut with CW speed
Replace the "practice mode on/off" on a long press
of the tune button, with setting the CW speed. This shortcut appears to
be much more useful.

Also removed the functionality of doing CW with the PTT key
and added the display of the current CW speed to the
main menu.